Fixed the source issue so the webpage opens when you do /RSVP in the URL. Made the RSVP form more organized.
