+++ views/js/evently/src/App.js
@@ -11,7 +11,7 @@ import CreateEvent from './CreateEvent.jsx';
 import CommunityPage from './CommunityPage.jsx';
 import AboutUs from './AboutUs.jsx';
 import RSVPForm from "./RSVPForm.jsx";
-import RSVPButton from "./RSVPButton/RSVPButton.jsx";
+import RSVPButton from "../src/RSVPButton.jsx";
 
 function App() {
   return (
