+++ views/js/evently/src/RSVPButton.jsx
@@ -3,22 +3,17 @@ import './RSVPButton.css'
     
 function RSVPButton() {
     return (
-     <div>
-            <head>
-                <meta charset="UTF-8" />
-                <meta name="author" content="Vid&Safiya" />
-                <meta name="description" content="This page creates the RVSP Button"/>
-                <title>RVSP</title>
-                <link rel="icon" href="html5.png" type="image/x-icon"/>
-                <link rel="stylesheet" href="RSVPButton.css" type="text/css"/>
-            </head>
-
-     <body>
-        <h1>"Event Name: ___"</h1> 
+     <div className="rsvp-event">
+       
+    
+        {/*<h1>"Event Name: ___"</h1> 
         <h2> "Location:___"</h2>
-        <h2> "Description:___"</h2>  {/* Is this the same description as the one the user is suppose to enter? It might be too long for the design */}
+        <h2> "Description:___"</h2>  {/*Is this the same description as the one the user is suppose to enter? It might be too long for the design */}
 
         <div className="UserInfo">
+          <h1>"Event Name: ___"</h1> 
+          <h2> "Location:___"</h2>
+          <h2> "Description:___"</h2> 
           <form action="action_page.php">
             <div className="row">
               <div className="col-25">
@@ -27,6 +22,7 @@ function RSVPButton() {
               <div className="col-75">
                 <input type="text" id="Email" name="Email" placeholder="Enter your email..." />
               </div>
+              <br />
             </div>
             <div className="row">
               <div className="col-25">
@@ -35,14 +31,16 @@ function RSVPButton() {
               <div className="col-75">
                 <input type="text" id="phoneNum" name="Phone #" placeholder="Enter your phone number..." />
               </div>
+              <br />
             </div>
           </form>
+          <br />
+          <div className="button">
+            <p>RSVP Here!</p>
+          </div>
         </div>
 
-        <div className="button">
-            <p>RSVP Here!</p>
-        </div>
-     </body>
+     
     </div>
  )
 };
